requirements.txt
Flask==0.8
Jinja2==2.6
Paver==1.0.5
Pillow==1.7.5
Werkzeug==0.8.2
distribute==0.6.24
gevent==0.13.6
greenlet==0.3.2
gunicorn==0.13.4
-e git://github.com/Greplin/scales.git#egg=scales
-e git://github.com/NorthIsUp/singletonmixin.git#egg=singletonmixin
wsgiref==0.1.2
raven2==1.0.0
blinker
httplib2
redis
Flask-Cache
